flask/requirements.txt
numpy
pillow
matplotlib
flask
opencv-python
rawpy
boto3


